[ks44team04] 결제상세페이지 테이블 병합, Mapper와 dto수정 author: 김석환 file: paymentDetail.html, PaymentHistory.java, PaymentInfo.java, paymentList.html, PaymentMapper.xml, PaymentTotal.java detail: 결제상세페이지 테이블을 병합하고 Mapper와 dto를 수정/삭제하였다.
