+++ ks44team04/src/main/resources/templates/admin/paymentDetail.html
@@ -51,33 +51,41 @@ th, td {
 					<thead>
 						<tr>
 							<th>주문번호</th>
-							<th>주문상세번호</th>
 							<th style="width: 100px">회원아이디</th>
+							<th>주문상세번호</th>
 							<th>상품정보</th>
 							<th style="width: 80px">결제수단</th>
 							<th style="width: 80px">영업사</th>
 							<th>카드번호/환불계좌</th>
 							<th>결제금액</th>
-							<th>결제일자</th>
+							<th>승인일시/입금일시</th>
 							<th>관리</th>
 						</tr>
 					</thead>
 					<tbody>
-						<tr th:if="${not #lists.isEmpty(paymentDetailList)}" th:each="p : ${paymentDetailList}">
-							<td th:text="${p.paymentInfo.orderNum}"></td>
-							<td th:text="${p.orderDetailCode}"></td>
-							<td th:text="${p.paymentInfo.buyerId}"></td>
-							<td th:text="${p.goodsCode}"></td>
-							<td th:text="${p.paymentMethod}"></td>
-							<td th:text="${p.cardCompany}"></td>
-							<td th:text="${p.cardNumber}"></td>
-							<td th:text="${'₩' + #numbers.formatInteger(p.approvePrice, 3, 'COMMA')}"></td>
-							<td th:text="${p.approveDate}"></td>
-							<td>
-								<a href="#"><i class="fa fa-edit"></i> 수정</a>
-								<a href="#"><i class="fa fa-close"></i>삭제</a>
-							</td>
-						</tr>
+						<th:block th:if="${not #lists.isEmpty(paymentDetailList)}" th:each="p : ${paymentDetailList}">
+							<th:block th:if="${not #lists.isEmpty(p.paymentTotal)}" th:each="pt, status : ${p.paymentTotal}">
+								<tr>
+									<th:block th:if="${status.index == 0}">				
+										<td th:rowspan="${status.size}" th:text="${p.orderNum}"></td>
+										<td th:rowspan="${status.size}" th:text="${p.buyerId}"></td>
+									</th:block>	
+									<td th:text="${pt.orderDetailCode}"></td>
+									<td th:text="${pt.goodsName}"></td>
+									<td th:text="${pt.paymentMethod}"></td>
+									<td th:text="${pt.cardCompany}"></td>
+									<td th:text="${pt.cardNumber}"></td>
+									<td th:text="${'₩' + #numbers.formatInteger(pt.approvePrice, 3, 'COMMA')}"></td>
+									<th:block th:if="${status.index == 0}">
+										<td th:rowspan="${status.size}" th:text="${p.approveDate}"></td>
+									</th:block>
+									<td>
+										<a href="#"><i class="fa fa-edit"></i> 수정</a>
+										<a href="#"><i class="fa fa-close"></i>삭제</a>
+									</td>
+								</tr>
+							</th:block>
+						</th:block>
 						<tr th:unless="${not #lists.isEmpty(paymentDetailList)}">
 							<td colspan="10">등록된 상품이 없습니다.</td>
 						</tr>
